requirements.txt
tweepy
tensorflow
mysql-connector-python